Move functions into util files and import
