+++ src/Components/MovieDisplay/MovieDisplay.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from "react";
 import { fetchMovies } from "../../API/apiCalls";
 import MovieThumb from "../MovieThumb/MovieThumb";
 import "./MovieDisplay.scss";
+import { filterMovieResults, cleanMovieData } from "../../utilities";
 
 const MovieDisplay = () => {
   const [movies, setMovies] = useState([]);
@@ -13,30 +14,6 @@ const MovieDisplay = () => {
       setMovies(clean);
     });
   }, []);
-
-  const filterMovieResults = (rawData) => {
-    return rawData.results.filter((result) => {
-      return (
-        result.type === "Movie" &&
-        result.title.includes("Boku") &&
-        result.start_date
-      );
-    });
-  };
-
-  const cleanMovieData = (dirtyData) => {
-    return dirtyData.map((data) => {
-      return {
-        image_url: data.image_url,
-        rated: data.rated,
-        start_date: data.start_date,
-        synopsis: data.synopsis,
-        url: data.url,
-        title: data.title,
-        mal_id: data.mal_id,
-      };
-    });
-  };
 
   const generateMovieThumbs = () => {
     return movies.map((movie, index) => {
